pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-speech-to-text-server"
version = "0.1.0"
description = "MCP Server for Speech-to-Text using Whisper and FastMCP with VAD"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch",
    "sounddevice",
    "numpy",
    "scipy",
    "openai-whisper",
    "fastmcp",
    "webrtcvad"
]

[project.scripts]
mcp-speech-to-text-server = "mcp_speech_to_text_server.main:mcp.run"

[tool.setuptools]
packages = ["mcp_speech_to_text_server"] 